edit entity

+++ src/main/java/com/example/demo/Service/impl/CategoryServiceImpl/ProductServiceImpl.java
@@ -14,7 +14,7 @@ public interface ProductServiceImpl {
 
     int totalItem();
 
-    Optional<ProductsDtoOut> getById(int id);
+    Optional<ProductsDtoOut> getById(Long id);
 
     List<ProductsDtoOut> getBestSeller();
